solve-cart-only -accept-one-store-product

+++ src/modules/orders/v1/dashboard/orders.controller.ts
@@ -1,3 +1,4 @@
+import { StoreController } from "./../../../stores/v1/website/stores.controller";
 import { Op } from "sequelize";
 import { AppError } from "../../../../utils/appError";
 import { ICart, IOrder } from "../../../../utils/shared.types";
@@ -60,7 +61,15 @@ export class OrderController {
 
   public async delete(req: Request) {
     const { id } = req.params;
-
+    const storeId = req.user?.storeId;
+    const order = (
+      await this.service.findOneByIdOrThrowError(id, {
+        attributes: ["id", "storeId"],
+      })
+    ).toJSON() as IOrder;
+    if (req.user?.role !== "superAdmin" && order.storeId !== storeId) {
+      throw new AppError("forbiden", 403);
+    }
     // Delete the order
     return this.service.delete(id);
   }
@@ -81,7 +90,7 @@ export class OrderController {
       include: [
         {
           model: OrderItem,
-          attributes: ["id", "quantity", "price", "cartId", "skuId"],
+          attributes: ["id", "quantity", "price"],
           include: [
             {
               model: Product,
@@ -98,39 +107,34 @@ export class OrderController {
 
   public async getAll(req: Request) {
     // Calculate offset for pagination
-
+    let storeId = req.user?.storeId;
     const { limit, offset, order, orderBy } = handlePaginationSort(req.query);
     let { search } = req.query;
     this.service.validateGetAllQuery({ search });
     const options: any = {
       attributes: {
-        exclude: [
-          "deletedAt",
-          "updatedAt",
-          "status",
-          "paymentStatus",
-          "storeId",
-        ],
+        exclude: ["deletedAt", "updatedAt", "status", "paymentStatus"],
       },
       offset,
       limit,
+      where: {},
       order: [[orderBy, order]],
 
-      include: [
-        {
-          model: OrderItem,
-          attributes: ["id", "quantity", "price", "cartId", "skuId"],
-          include: [
-            {
-              model: Product,
-              attributes: ["id", "name", "description", "image"],
-            },
-            { model: ProductSku, attributes: ["id", "price", "sku"] },
-          ],
-        },
-      ],
+      // include: [
+      //   {
+      //     model: OrderItem,
+      //     attributes: ["id", "quantity", "price", "cartId", "skuId"],
+      //     include: [
+      //       {
+      //         model: Product,
+      //         attributes: ["id", "name", "description", "image"],
+      //       },
+      //       { model: ProductSku, attributes: ["id", "price", "sku"] },
+      //     ],
+      //   },
+      // ],
     };
-
+    if (storeId) options.where.storeId = storeId;
     const date = await this.service.getAll(options);
 
     return date;
